+++ graphqlapi/graphql/execute.lua
@@ -193,11 +193,7 @@ local function completeValue(fieldType, result, subSelections, context, opts)
   end
 
   if result == nil then
-    if result then
-      return box.NULL
-    else
-      return nil
-    end
+      return result
   end
 
   if fieldTypeName == 'List' then