+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - `Add operations.get_operation_fields() to extract operation requested fields`
 - `Add utils.find() helper`
 - `Use cmake build type instead builtin`
+- `Add cluster.get_candidates() method`
 
 ## 0.0.8